chore: move custom task to dedicated dir
